# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Weekly Commitments API
  description: |
    API for the Weekly Commitments module – plan creation, lifecycle management,
    reconciliation, manager review, RCDO integration, and AI-assisted suggestions.
  version: 1.0.0
  contact:
    name: ST6 Engineering
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
  /weeks/{weekStart}/plans:
    post:
      operationId: createPlan
      summary: Create a weekly plan
      description: |
        Create a plan for the specified week. Idempotent: returns 201 on create,
        200 if the plan already exists for this user/week.
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/WeekStartPath'
      responses:
        '200':
          description: Plan already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '201':
          description: Plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /weeks/{weekStart}/plans/me:
    get:
      operationId: getMyPlan
      summary: Get my plan for a given week
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/WeekStartPath'
      responses:
        '200':
          description: Plan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /plans/{planId}:
    get:
      operationId: getPlan
      summary: Get plan by ID
      description: Accessible by the plan owner or their manager.
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
      responses:
        '200':
          description: Plan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /weeks/{weekStart}/plans/{userId}:
    get:
      operationId: getUserPlan
      summary: Get a user's plan (manager-only)
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/WeekStartPath'
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /plans/{planId}/lock:
    post:
      operationId: lockPlan
      summary: Lock a weekly plan
      description: |
        Transitions a plan from DRAFT to LOCKED. Validates all commits,
        populates RCDO snapshots, and enforces chess rules.
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Plan locked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /plans/{planId}/start-reconciliation:
    post:
      operationId: startReconciliation
      summary: Start reconciliation
      description: |
        Transitions a plan from LOCKED to RECONCILING. If the plan is in
        DRAFT and the week has passed, performs a late lock first.
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Reconciliation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /plans/{planId}/submit-reconciliation:
    post:
      operationId: submitReconciliation
      summary: Submit reconciliation
      description: |
        Transitions from RECONCILING to RECONCILED. Requires all commits
        to have completionStatus and deltaReason for non-DONE items.
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Reconciliation submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /plans/{planId}/carry-forward:
    post:
      operationId: carryForward
      summary: Carry forward incomplete commits
      description: |
        Copies selected commits into the next week's plan (DRAFT).
        Transitions the current plan to CARRY_FORWARD.
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarryForwardRequest'
      responses:
        '200':
          description: Carry-forward executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /plans/{planId}/review:
    post:
      operationId: createReview
      summary: Submit manager review
      description: |
        Manager approves or requests changes to a plan. Request changes
        is blocked if carry-forward has already been executed.
      tags: [Manager]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerReview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /plans/draft-from-history:
    post:
      operationId: draftFromHistory
      summary: '"Start My Week" — create a draft plan pre-filled from history'
      description: |
        Creates (or updates) a DRAFT plan for the requested week, pre-filled with
        AI-suggested commits derived from the user's last 4 weeks of planning history.

        Suggestion sources:
        - CARRIED_FORWARD — commits not completed (status != DONE) in the most recent
          reconciled plan
        - RECURRING — commit titles or outcome IDs appearing in 2+ consecutive historical weeks
        - COVERAGE_GAP — reserved for future NextWorkService integration

        Each suggested commit is persisted in the new/existing DRAFT plan with its source
        stored in the tags[] array as "draft_source:<SOURCE>".

        Returns 409 if the target week already has a plan in a non-DRAFT state.
      tags: [Plans, AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftFromHistoryRequest'
      responses:
        '200':
          description: Draft plan created or updated with suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftFromHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WeekStartPath:
      name: weekStart
      in: path
      required: true
      description: ISO date of the Monday that starts the week (e.g., 2026-03-09)
      schema:
        type: string
        format: date
    PlanIdPath:
      name: planId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Optimistic lock version number
      schema:
        type: integer
  schemas:
    PlanState:
      type: string
      enum: [DRAFT, LOCKED, RECONCILING, RECONCILED, CARRY_FORWARD]
    ReviewStatus:
      type: string
      enum: [REVIEW_NOT_APPLICABLE, REVIEW_PENDING, CHANGES_REQUESTED, APPROVED]
    ChessPriority:
      type: string
      enum: [KING, QUEEN, ROOK, BISHOP, KNIGHT, PAWN]
...
    LockType:
      type: string
      enum: [ON_TIME, LATE_LOCK]
    CommitCategory:
      type: string
      enum: [DELIVERY, OPERATIONS, CUSTOMER, GTM, PEOPLE, LEARNING, TECH_DEBT]
    WeeklyPlan:
      type: object
      required: [id, orgId, ownerUserId, weekStartDate, state, reviewStatus, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        weekStartDate:
          type: string
          format: date
        state:
          $ref: '#/components/schemas/PlanState'
        reviewStatus:
          $ref: '#/components/schemas/ReviewStatus'
        lockType:
          $ref: '#/components/schemas/LockType'
        lockedAt:
          type: string
          format: date-time
          nullable: true
        carryForwardExecutedAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ManagerReview:
      type: object
      required: [id, weeklyPlanId, reviewerUserId, decision, comments, createdAt]
      properties:
        id:
          type: string
          format: uuid
        weeklyPlanId:
          type: string
          format: uuid
        reviewerUserId:
          type: string
          format: uuid
        decision:
          type: string
          enum: [APPROVED, CHANGES_REQUESTED]
        comments:
          type: string
        createdAt:
          type: string
          format: date-time
    CarryForwardRequest:
      type: object
      required: [commitIds]
      properties:
        commitIds:
          type: array
          items:
            type: string
            format: uuid
...
    CreateReviewRequest:
      type: object
      required: [decision, comments]
      properties:
        decision:
          type: string
          enum: [APPROVED, CHANGES_REQUESTED]
        comments:
          type: string
...
    CommitSource:
      type: string
      enum: [CARRIED_FORWARD, RECURRING, COVERAGE_GAP]
      description: |
        Source classification for AI-suggested commits produced by the
        draft-from-history service:
        - CARRIED_FORWARD: not completed in the most recent reconciled plan
        - RECURRING: similar title or outcome ID appeared in 2+ consecutive weeks
        - COVERAGE_GAP: coverage gap (future NextWorkService integration)
    SuggestedCommit:
      type: object
      required: [commitId, title, source]
      properties:
        commitId:
          type: string
          format: uuid
          description: ID of the newly created DRAFT commit
        title:
          type: string
        description:
          type: string
          nullable: true
        chessPriority:
          allOf:
            - $ref: '#/components/schemas/ChessPriority'
          nullable: true
        category:
          allOf:
            - $ref: '#/components/schemas/CommitCategory'
          nullable: true
        outcomeId:
          type: string
          format: uuid
          nullable: true
        nonStrategicReason:
          type: string
          nullable: true
        expectedResult:
          type: string
          nullable: true
        source:
          $ref: '#/components/schemas/CommitSource'
    DraftFromHistoryRequest:
      type: object
      required: [weekStart]
      properties:
        weekStart:
          type: string
          format: date
          description: |
            ISO date of the Monday that starts the target week (e.g. 2026-03-16).
            Must be a Monday; returns 422 otherwise.
    DraftFromHistoryResponse:
      type: object
      required: [planId, suggestedCommits]
      properties:
        planId:
          type: string
          format: uuid
          description: ID of the created or updated DRAFT plan
        suggestedCommits:
          type: array
          items:
            $ref: '#/components/schemas/SuggestedCommit'
          description: Ordered list of suggested commits added to the draft plan
...
    ApiError:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Valid JWT but insufficient access
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    Conflict:
      description: State conflict (version mismatch, invalid transition)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
...